Update etcd api; fix panics
